Group fails to list members if there is more than 100 #779
Added button to load more members in Group statistics.
All members are loaded if the total number of members are lower than (BATCH_SIZE + 10), otherwise only BATCH_SIZE will be loaded automatically. This is done to prevent a small amount of items to remain unloaded.
Fixed report group visible for non-admin users in groups and roles.